refactor(desktop): move providers domain modules to features/providers
providers, provider-pricing, routing, slot-routing moved.
all importers updated; vi.mock and dynamic imports in store tests fixed.
