+++ apps/desktop/src/__tests__/snapshots/empty-error-states.test.tsx
@@ -83,7 +83,7 @@ vi.mock('../../routing', () => ({
 
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { cleanup, render } from '@testing-library/react';
-import type { ProviderInfo } from '../../providers';
+import type { ProviderInfo } from '../../features/providers/providers';
 import type { AppState, AppActions } from '../../store/store';
 import type { Session, SessionId, WorkspaceId } from '@kay-am/types';
 import { useAppStore } from '../../store';
