+++ apps/desktop/src/components/NewSessionDialog.tsx
@@ -23,7 +23,7 @@ import type {
 } from '@kay-am/types';
 import { PROVIDER_CAPABILITIES, getDefaultTurnModel } from '@kay-am/core';
 import { shortModel } from '../agent-row-format';
-import { PROVIDER_LABEL_LOWER } from '../providers';
+import { PROVIDER_LABEL_LOWER } from '../features/providers/providers';
 import { settingBranchPrefix, DEFAULT_BRANCH_PREFIX } from '../settings';
 import { SESSION_FEATURES } from '../shared/lib/features';
 import { EMPTY_ARRAY, useAppStore } from '../store';
