+++ apps/desktop/src/components/chat/AuthRequiredCallout.tsx
@@ -1,7 +1,7 @@
 import { invoke } from '@tauri-apps/api/core';
 import { Button } from '@kay-am/ui';
 import type { ProviderId } from '@kay-am/types';
-import { PROVIDER_LABEL_LOWER } from '../../providers';
+import { PROVIDER_LABEL_LOWER } from '../../features/providers/providers';
 
 async function providerAction(id: ProviderId, action: 'login' | 'logout'): Promise<void> {
   return invoke('provider_action', { id, action });
