+++ apps/desktop/src/store/store.audit-retry.test.ts
@@ -83,7 +83,7 @@ vi.mock('@kay-am/db', () => ({
   clearAllNotifications: vi.fn(async () => undefined),
 }));
 
-vi.mock('../providers', () => ({
+vi.mock('../features/providers/providers', () => ({
   buildProviderList: () => [{ id: 'anthropic', binary: 'claude', connection: 'connected' }],
   checkProviderAuth: vi.fn(),
   getCursorStatus: vi.fn(),
@@ -91,7 +91,7 @@ vi.mock('../providers', () => ({
   getProviderStatus: vi.fn(),
 }));
 
-vi.mock('../routing', () => ({
+vi.mock('../features/providers/routing', () => ({
   resolveProviderForTurn: vi.fn(async () => ({
     selectedProvider: 'anthropic',
     selectedModel: 'claude-3-5-sonnet-latest',
@@ -136,7 +136,7 @@ vi.mock('../shared/lib/repo', () => ({
   validateGitRepo: vi.fn(),
 }));
 
-vi.mock('../provider-pricing', () => ({
+vi.mock('../features/providers/provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
@@ -330,7 +330,7 @@ describe('audit retry queue — drain worker (happy path)', () => {
     (getSetting as ReturnType<typeof vi.fn>).mockResolvedValue(null);
 
     const { getProviderStatus, getCursorStatus, getCodexStatus, checkProviderAuth } =
-      await import('../providers');
+      await import('../features/providers/providers');
     (getProviderStatus as ReturnType<typeof vi.fn>).mockResolvedValue({
       state: 'connected',
       identity: 'test',
