--- apps/desktop/src/utils/display-path.ts
@@ -1,5 +1,0 @@
-export function displayPath(absPath: string, workingDir: string | null | undefined): string {
-  if (!workingDir) return absPath;
-  const root = workingDir.endsWith('/') ? workingDir : `${workingDir}/`;
-  return absPath.startsWith(root) ? absPath.slice(root.length) : absPath;
-}
